Refactor code forsignup component
