+++ src/components/SignUpForm/SignUpMain.tsx
@@ -1,114 +1,23 @@
-import React, {useState, useEffect} from 'react'
-import { useNavigate } from 'react-router-dom';
-import {Box, Backdrop, CircularProgress} from '@mui/material';
 
+import {Box, Backdrop, CircularProgress, TextField, Typography, Button} from '@mui/material';
+import { StyledCancelButton, StyledLoginButton } from './Style';
 
-import SignUpForm from './SignUpForm'
 import { StyledSignUpMainContainer } from './Style'
-import { useAppDispatch, useAppSelector } from '../../hooks/storeHooks';
-import { signUp as postSignUpDetails, signUpActions } from '../../features/auths/signUp/signUpSlice';
-import { userActions } from '../../features/auths/user/userSlice';
-import { msgAction } from '../../features/msgHandler/msgHandler';
-import { getAllTrips } from '../../features/SavedTrip/SavedTrip';
+
+interface SignUpFormProps {
+  handleSignUpOnChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  handleClose: () => void;
+  signUpButtonDisabled: boolean;
+  handleSignUp: () => void;
+  handleNavigateToLogin: () => void;
+  signUpData: { email: string; password: string; confirmPassword: string};
+  openBackDrop: boolean;
+}
 
 
-const SignUpMain = () => {
-  const navigate = useNavigate()
-  const dispatch = useAppDispatch()
-  const { signUp, user: {user} } = useAppSelector(state => state)
 
-  const [openBackDrop, setOpenBackDrop] = React.useState(false);
-  
-  const [signUpData, setSignUpData] = useState({
-    email: "",
-    password: "",
-    confirmPassword: ""
-  })
-
-  useEffect(() => {
-    
-    if (user){
-      dispatch(signUpActions.resetSignUp())
-      dispatch(getAllTrips({token: user.token}))
-    } 
-  }, [user, dispatch])
-
-  useEffect(() => {
-   
-    
-    if(signUp.isLoading){
-      setOpenBackDrop(true)
-    } else {
-      setOpenBackDrop(false)
-    }
-    
-    if (signUp.data && signUp.data.status === 'success'){
-      
-      dispatch(userActions.setUser(signUp.data))
-      dispatch(msgAction.getSuccessMsg("Account created successfully"))
-      
-      navigate(-1)
-      return
-    }
-    if (signUp.error) {
-      dispatch(msgAction.getErrorMsg("Account already exist!"))
-      dispatch(signUpActions.resetSignUp())
-      return
-    }
-  }, [signUp, dispatch, navigate])
-
-
-  const [signUpButtonDisabled, setSignUpButtonDisabled] = useState(true);
-
-  const handleSignUpOnChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    event.preventDefault();
-
-    if (signUpData.email && signUpData.password && signUpData.confirmPassword) {
-      setSignUpButtonDisabled(false);
-    } else {
-      setSignUpButtonDisabled(true);
-    }
-
-    setSignUpData({
-      ...signUpData,
-      [event.target.id]: event.target.value
-    })
-  }
-
-  const handleClose = () => {
-    setSignUpButtonDisabled(true)
-    setSignUpData({
-      email: "",
-      password: "",
-      confirmPassword: ""
-    })
-    navigate(-1)
-  };
-
-  const handleSignUp = () => {
-    const {email, password, confirmPassword } = signUpData
-    if (email && password && confirmPassword && password !== confirmPassword){
-      dispatch(msgAction.getErrorMsg("Password mismatch"))
-      return;
-    }
-
-    if (email && password && confirmPassword && password === confirmPassword) {
-      dispatch(postSignUpDetails({email, password}))
-
-      setSignUpData({
-        email: "",
-        password: "",
-        confirmPassword: ""
-      })
-    }
-
-  }
-
-  const handleNavigateToLogin = () => {    
-     navigate("/Login")
-  }
-
-
+const SignUpMain = ({handleSignUpOnChange, handleClose, signUpButtonDisabled, handleSignUp, handleNavigateToLogin, signUpData, openBackDrop}: SignUpFormProps) => {
+ 
   return (
     <Box>      
       <Box>
@@ -120,18 +29,99 @@ const SignUpMain = () => {
         </Backdrop>
       </Box>      
 
-      <Box>
-        <StyledSignUpMainContainer>
-          <SignUpForm 
-            handleSignUpOnChange={handleSignUpOnChange}
-            handleClose={handleClose}
-            signUpButtonDisabled={signUpButtonDisabled}
-            handleSignUp={handleSignUp}
-            handleNavigateToLogin={handleNavigateToLogin}
-            signUpData={signUpData}
-          />
-        </StyledSignUpMainContainer>
-      </Box>
+      <StyledSignUpMainContainer>
+        <Box>
+          <Box
+            component="form"
+            sx={{
+              
+              display: "flex",
+              flexDirection: "column",
+              justifyContent: "center",
+              alignItems: "center",
+              gap: "1rem"
+            }}
+          >
+            <TextField
+              autoFocus
+              margin="dense"
+              id="email"
+              label="Email Address"
+              type="email"
+              fullWidth
+              variant="filled"
+              value={signUpData.email}
+              onChange={handleSignUpOnChange}
+              sx={{
+                backgroundColor: "white"
+              }}
+            />
+
+            <TextField            
+              margin="dense"
+              id="password"
+              label="Password"
+              type="password"
+              autoComplete="current-password"
+              fullWidth
+              variant="filled"
+              value={signUpData.password}
+              onChange={handleSignUpOnChange}
+              sx={{
+                backgroundColor: "white"
+              }}
+            />
+
+            <TextField            
+              margin="dense"
+              id="confirmPassword"
+              label="Re-enter Password"
+              type="password"
+              autoComplete="password"
+              fullWidth
+              variant="filled"
+              value={signUpData.confirmPassword}
+              onChange={handleSignUpOnChange}
+              sx={{
+                backgroundColor: "white"
+              }}
+            />
+            
+          </Box>
+          <Box
+            sx={{
+              display: "flex",
+              justifyContent: "center",
+              alignItems: "center",
+              gap: "2rem",
+              margin: "2rem"
+
+            }}    
+          >
+            <StyledCancelButton onClick={handleClose}>CANCEL</StyledCancelButton>
+            <StyledLoginButton disabled={signUpButtonDisabled} onClick={handleSignUp}>SIGN UP</StyledLoginButton>
+          </Box>
+
+          <Box
+            sx={{
+              display: "flex",
+              flexDirection: "column",
+              justifyContent: "center",
+              alignItems: "center",
+              gap: "1rem",
+              margin: "2rem"
+            }}
+          >
+            <Typography variant="subtitle1">
+              Already have an account?
+            </Typography>
+
+            <Button variant="text" onClick={handleNavigateToLogin}>          
+              Login
+            </Button>
+          </Box>                    
+        </Box>
+      </StyledSignUpMainContainer>
     </Box>
   )
 }
